Fix triggering method bottomReached
When there are only one coin transactions (for ex. only BTC) and user switches filter from All to BTC it may doesn't load the next page.
The reason is that the logic for loading next page was in onBindViewHolder.
This method may be not called if the new transaction content is the same as old one.