Reset list state instead of scrolling to top
There is a task to show the list from the beginning when the sort type or account is changed.
After some research I came to the solution with list state.
The list state initiated with the first item equal to 0.
Reiniting does the same.
Instead of passing the parameter for scrolling to top from the viewModel to view I made the list state dependant on the account and sort type.
It allows to hold this logic completely inside of view.